Spostato il metodo updateMotion nelle classi figlie. Aggiunta la classe Bullet, in fase di lavorazione.
