+++ core/src/com/philosfight/game/game/objects/AbstractGameObject.java
@@ -59,41 +59,7 @@ public abstract class AbstractGameObject {
         position.y += velocity.y * deltaTime;
     }
 
-    protected void updateMotionX (float deltaTime) {
-        if (velocity.x != 0) {
-            // Apply friction
-            if (velocity.x > 0) {
-                velocity.x =
-                        Math.max(velocity.x - friction.x * deltaTime, 0);
-            } else {
-                velocity.x =
-                        Math.min(velocity.x + friction.x * deltaTime, 0);
-            }
-        }
-        // Apply acceleration
-        velocity.x += acceleration.x * deltaTime;
-        // Make sure the object's velocity does not exceed the
-        // positive or negative terminal velocity
-        velocity.x = MathUtils.clamp(velocity.x,
-                -terminalVelocity.x, terminalVelocity.x);
-    }
-
-    protected void updateMotionY (float deltaTime) {
-        if (velocity.y != 0) {
-            // Apply friction
-            if (velocity.y > 0) {
-                velocity.y = Math.max(velocity.y - friction.y * deltaTime, 0);
-            } else {
-                velocity.y = Math.min(velocity.y + friction.y * deltaTime, 0);
-            }
-        }
-        // Apply acceleration
-        velocity.y += acceleration.y * deltaTime;
-        // Make sure the object's velocity does not exceed the
-        // positive or negative terminal velocity
-        velocity.y = MathUtils.clamp(velocity.y, -terminalVelocity.y, terminalVelocity.y);
-    }
-
-
+    protected void updateMotionX (float deltaTime){};
+    protected void updateMotionY (float deltaTime){};
     public abstract void render (SpriteBatch batch);
 }
