+++ core/src/com/philosfight/game/utils/Constants.java
@@ -1,15 +1,15 @@
 package com.philosfight.game.utils;
 
 public class Constants {
-//    Grandezza visuale del giocatore
+    //Grandezza visuale del giocatore
     public static final float VIEWPORT_WIDTH = 5.0f;
     public static final float VIEWPORT_HEIGHT = 5.0f;
-    //    Posizione del file atlas che contiene le informazioni sulle texture dell'arena
+    //Posizione del file atlas che contiene le informazioni sulle texture dell'arena
     public static final String TEXTURE_ATLAS_ARENA  = "Images/pack_arena.atlas";
-//    Posizione del file immagine dell'arena
+    //Posizione del file immagine dell'arena
     public static final String ARENA_00 = "Arenas/arena00.png";
-//        GUI Width
+    //GUI Width
     public static final float VIEWPOER_GUI_WIDTH = 800.0f;
-//    GUI Height
+    //GUI Height
     public static final float VIEWPOER_GUI_HEIGHT = 400.0f;
 }
