+++ desktop/desktop.iml
@@ -44,6 +44,12 @@
     <orderEntry type="library" exported="" name="Gradle: lwjgl-platform-2.9.2-natives-osx" level="project" />
     <orderEntry type="library" exported="" name="Gradle: jinput-2.0.5" level="project" />
     <orderEntry type="library" exported="" name="Gradle: jutils-1.0.0" level="project" />
+    <orderEntry type="library" exported="" scope="RUNTIME" name="Gradle: box2dlights-1.4" level="project" />
+    <orderEntry type="library" exported="" scope="RUNTIME" name="Gradle: gdx-box2d-1.9.9" level="project" />
+    <orderEntry type="library" exported="" scope="RUNTIME" name="Gradle: gdx-bullet-1.9.9" level="project" />
+    <orderEntry type="library" exported="" scope="RUNTIME" name="Gradle: jinput-platform-2.0.5-natives-linux" level="project" />
+    <orderEntry type="library" exported="" scope="RUNTIME" name="Gradle: jinput-platform-2.0.5-natives-windows" level="project" />
+    <orderEntry type="library" exported="" scope="RUNTIME" name="Gradle: jinput-platform-2.0.5-natives-osx" level="project" />
     <orderEntry type="library" exported="" scope="TEST" name="Gradle: box2dlights-1.4" level="project" />
     <orderEntry type="library" exported="" scope="TEST" name="Gradle: gdx-box2d-1.9.9" level="project" />
     <orderEntry type="library" exported="" scope="TEST" name="Gradle: gdx-bullet-1.9.9" level="project" />
