+++ local.properties
@@ -4,5 +4,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Sun Sep 01 15:50:44 CEST 2019
-sdk.dir=C\:\\Users\\allez\\AppData\\Local\\Android\\Sdk
+#Sun Sep 01 19:35:45 CEST 2019
+sdk.dir=C\:\\Users\\jacks\\AppData\\Local\\Android\\Sdk
